+++ camunda-formio-tasklist-vue/src/services/CamundaRest.ts
@@ -1,5 +1,5 @@
 import {
-  bpmAxios, bpmAxiosWithHal, engine, formExt,
+  bpmAxios, bpmAxiosGet, bpmAxiosWithHal, engine,
 } from "../services";
 
 export const CamundaRest = {
@@ -274,7 +274,7 @@ export const CamundaRest = {
   },
 
   async listForms (bearerToken: string, CamundaUrl: string) {
-    return await bpmAxios(bearerToken, CamundaUrl).get(`/${formExt}/form`);
+    return await bpmAxiosGet(bearerToken, CamundaUrl).get(`/form`);
   },
 
   async getProcessDiagramXML (